requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.1
django-model-utils==3.0.0
django-widget-tweaks
facebook-sdk==2.0.0
gunicorn==19.6.0
idna==2.6
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.2
